requirements.txt
torch
torchvision
diffusers
transformers
accelerate
xformers
Pillow
requests
numpy
runpod